+++ src/main/java/subway/presentation/support/ExceptionAdvice.java
@@ -1,9 +1,8 @@
-package subway.presentation.advice;
+package subway.presentation.support;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.dao.DataAccessException;
-import org.springframework.dao.DuplicateKeyException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
